Add tests and refactored command
