+++ src/main/java/seedu/address/model/person/Remark.java
@@ -1,0 +1,4 @@
+package seedu.address.model.person;
+
+public class Remark {
+}
