+++ src/test/java/seedu/address/commons/core/Messages.java
@@ -1,0 +1,8 @@
+package seedu.address.commons.core;
+
+public class Messages {
+
+    public static final String MESSAGE_INVALID_COMMAND_FORMAT = "invalid command";
+
+    public static final String MESSAGE_UNKNOWN_COMMAND = "unknown command";
+}
